requirements.txt
google-cloud-storage==2.1.0
django-storages==1.12.3
django==3.2.7
djangorestframework==3.12.4
pandas==1.3.3
gunicorn==20.1.0
google-cloud-bigquery==2.23.2
djangorestframework-jwt==1.11.0
plotly==5.7.0
networkx==2.8
decorator==5.0.9
python-dotenv==0.20.0